pyproject.toml
[project]
name = "smallbench"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apropos-ai==0.1.12",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
